backend/requirements.txt
django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
drf-yasg==1.21.7
pytest-django==4.8.0
factory-boy==3.3.0
python-dateutil==2.9.0